+++ contracts/facets/verification/secp256r1/utils/LibSecp256r1.sol
@@ -18,21 +18,14 @@ struct JPoint {
 }
 
 library LibSecp256r1 {
-    uint256 constant gx =
-        0x6B17D1F2E12C4247F8BCE6E563A440F277037D812DEB33A0F4A13945D898C296;
-    uint256 constant gy =
-        0x4FE342E2FE1A7F9B8EE7EB4A7C0F9E162BCE33576B315ECECBB6406837BF51F5;
-    uint256 public constant pp =
-        0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFF;
+    uint256 constant gx = 0x6B17D1F2E12C4247F8BCE6E563A440F277037D812DEB33A0F4A13945D898C296;
+    uint256 constant gy = 0x4FE342E2FE1A7F9B8EE7EB4A7C0F9E162BCE33576B315ECECBB6406837BF51F5;
+    uint256 public constant pp = 0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFF;
 
-    uint256 public constant nn =
-        0xFFFFFFFF00000000FFFFFFFFFFFFFFFFBCE6FAADA7179E84F3B9CAC2FC632551;
-    uint256 constant a =
-        0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFC;
-    uint256 constant b =
-        0x5AC635D8AA3A93E7B3EBBD55769886BC651D06B0CC53B0F63BCE3C3E27D2604B;
-    uint256 constant MOST_SIGNIFICANT =
-        0xc000000000000000000000000000000000000000000000000000000000000000;
+    uint256 public constant nn = 0xFFFFFFFF00000000FFFFFFFFFFFFFFFFBCE6FAADA7179E84F3B9CAC2FC632551;
+    uint256 constant a = 0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFC;
+    uint256 constant b = 0x5AC635D8AA3A93E7B3EBBD55769886BC651D06B0CC53B0F63BCE3C3E27D2604B;
+    uint256 constant MOST_SIGNIFICANT = 0xc000000000000000000000000000000000000000000000000000000000000000;
 
     /*
      * Verify
@@ -42,12 +35,7 @@ library LibSecp256r1 {
      * @param S - signature half S
      * @param input - hashed message
      */
-    function Verify(
-        uint256[2] memory q,
-        uint r,
-        uint s,
-        uint e
-    ) internal view returns (bool) {
+    function Verify(uint256[2] memory q, uint256 r, uint256 s, uint256 e) internal view returns (bool) {
         if (q[0] > pp - 1 || q[1] > pp - 1) {
             return false;
         }
@@ -55,16 +43,8 @@ library LibSecp256r1 {
             return false;
         }
         if (
-            mulmod(q[1], q[1], pp) !=
-            addmod(
-                addmod(
-                    mulmod(q[0], mulmod(q[0], q[0], pp), pp),
-                    mulmod(a, q[0], pp),
-                    pp
-                ),
-                b,
-                pp
-            )
+            mulmod(q[1], q[1], pp)
+                != addmod(addmod(mulmod(q[0], mulmod(q[0], q[0], pp), pp), mulmod(a, q[0], pp), pp), b, pp)
         ) {
             return false;
         }
@@ -73,23 +53,22 @@ library LibSecp256r1 {
         return VerifyWithPrecompute(points, r, s, e);
     }
 
-    function VerifyWithPrecompute(
-        JPoint[16] memory points,
-        uint r,
-        uint s,
-        uint e
-    ) internal view returns (bool) {
+    function VerifyWithPrecompute(JPoint[16] memory points, uint256 r, uint256 s, uint256 e)
+        internal
+        view
+        returns (bool)
+    {
         if (r == 0 || s == 0 || r >= nn || s >= nn) {
             return false;
         }
 
-        uint w = _primemod(s, nn);
+        uint256 w = _primemod(s, nn);
 
-        uint u1 = mulmod(e, w, nn);
-        uint u2 = mulmod(r, w, nn);
+        uint256 u1 = mulmod(e, w, nn);
+        uint256 u2 = mulmod(r, w, nn);
 
-        uint x;
-        uint y;
+        uint256 x;
+        uint256 y;
 
         (x, y) = ShamirMultJacobian(points, u1, u2);
         return (x == r);
@@ -102,34 +81,25 @@ library LibSecp256r1 {
      * the individual points for a single pass are precomputed
      * overall this reduces the number of additions while keeping the same number of doublings
      */
-    function ShamirMultJacobian(
-        JPoint[16] memory points,
-        uint u1,
-        uint u2
-    ) internal view returns (uint, uint) {
-        uint x = 0;
-        uint y = 0;
-        uint z = 0;
-        uint bits = 128;
-        uint index = 0;
+    function ShamirMultJacobian(JPoint[16] memory points, uint256 u1, uint256 u2)
+        internal
+        view
+        returns (uint256, uint256)
+    {
+        uint256 x = 0;
+        uint256 y = 0;
+        uint256 z = 0;
+        uint256 bits = 128;
+        uint256 index = 0;
 
         while (bits > 0) {
             if (z > 0) {
                 (x, y, z) = _modifiedJacobianDouble(x, y, z);
                 (x, y, z) = _modifiedJacobianDouble(x, y, z);
             }
-            index =
-                ((u1 & MOST_SIGNIFICANT) >> 252) |
-                ((u2 & MOST_SIGNIFICANT) >> 254);
+            index = ((u1 & MOST_SIGNIFICANT) >> 252) | ((u2 & MOST_SIGNIFICANT) >> 254);
             if (index > 0) {
-                (x, y, z) = _jAdd(
-                    x,
-                    y,
-                    z,
-                    points[index].x,
-                    points[index].y,
-                    points[index].z
-                );
+                (x, y, z) = _jAdd(x, y, z, points[index].x, points[index].y, points[index].z);
             }
             u1 <<= 2;
             u2 <<= 2;
@@ -139,9 +109,7 @@ library LibSecp256r1 {
         return (x, y);
     }
 
-    function _preComputeJacobianPoints(
-        uint256[2] memory q
-    ) internal pure returns (JPoint[16] memory points) {
+    function _preComputeJacobianPoints(uint256[2] memory q) internal pure returns (JPoint[16] memory points) {
         points[0] = JPoint(0, 0, 0);
         points[1] = JPoint(q[0], q[1], 1); // u2
         points[2] = _jPointDouble(points[1]);
@@ -163,23 +131,18 @@ library LibSecp256r1 {
         points[15] = _jPointAdd(points[12], points[3]);
     }
 
-    function _jPointAdd(
-        JPoint memory p1,
-        JPoint memory p2
-    ) internal pure returns (JPoint memory) {
-        uint x;
-        uint y;
-        uint z;
+    function _jPointAdd(JPoint memory p1, JPoint memory p2) internal pure returns (JPoint memory) {
+        uint256 x;
+        uint256 y;
+        uint256 z;
         (x, y, z) = _jAdd(p1.x, p1.y, p1.z, p2.x, p2.y, p2.z);
         return JPoint(x, y, z);
     }
 
-    function _jPointDouble(
-        JPoint memory p
-    ) internal pure returns (JPoint memory) {
-        uint x;
-        uint y;
-        uint z;
+    function _jPointDouble(JPoint memory p) internal pure returns (JPoint memory) {
+        uint256 x;
+        uint256 y;
+        uint256 z;
         (x, y, z) = _modifiedJacobianDouble(p.x, p.y, p.z);
         return JPoint(x, y, z);
     }
@@ -188,17 +151,13 @@ library LibSecp256r1 {
      * @desription returns affine coordinates from a jacobian input follows
      * golang elliptic/crypto library
      */
-    function _affineFromJacobian(
-        uint x,
-        uint y,
-        uint z
-    ) internal view returns (uint ax, uint ay) {
+    function _affineFromJacobian(uint256 x, uint256 y, uint256 z) internal view returns (uint256 ax, uint256 ay) {
         if (z == 0) {
             return (0, 0);
         }
 
-        uint zinv = _primemod(z, pp);
-        uint zinvsq = mulmod(zinv, zinv, pp);
+        uint256 zinv = _primemod(z, pp);
+        uint256 zinvsq = mulmod(zinv, zinv, pp);
 
         ax = mulmod(x, zinvsq, pp);
         ay = mulmod(y, mulmod(zinvsq, zinv, pp), pp);
@@ -209,14 +168,11 @@ library LibSecp256r1 {
      * @description performs Jacobian addition as defined below:
      * http://www.hyperelliptic.org/EFD/g1p/data/shortw/jacobian/addition/add-2007-bl
      */
-    function _jAdd(
-        uint p1,
-        uint p2,
-        uint p3,
-        uint q1,
-        uint q2,
-        uint q3
-    ) internal pure returns (uint r1, uint r2, uint r3) {
+    function _jAdd(uint256 p1, uint256 p2, uint256 p3, uint256 q1, uint256 q2, uint256 q3)
+        internal
+        pure
+        returns (uint256 r1, uint256 r2, uint256 r3)
+    {
         if (p3 == 0) {
             r1 = q1;
             r2 = q2;
@@ -232,9 +188,7 @@ library LibSecp256r1 {
         }
 
         assembly {
-            let
-                pd
-            := 0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFF
+            let pd := 0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFF
             let z1z1 := mulmod(p3, p3, pd) // Z1Z1 = Z1^2
             let z2z2 := mulmod(q3, q3, pd) // Z2Z2 = Z2^2
 
@@ -246,47 +200,38 @@ library LibSecp256r1 {
 
             let p3q3 := addmod(p3, q3, pd)
 
-            if lt(u2, u1) {
-                u2 := add(pd, u2) // u2 = u2+pd
-            }
+            if lt(u2, u1) { u2 := add(pd, u2) } // u2 = u2+pd
+
             let h := sub(u2, u1) // H = U2-U1
 
             let i := mulmod(0x02, h, pd)
             i := mulmod(i, i, pd) // I = (2*H)^2
 
             let j := mulmod(h, i, pd) // J = H*I
-            if lt(s2, s1) {
-                s2 := add(pd, s2) // u2 = u2+pd
-            }
+            if lt(s2, s1) { s2 := add(pd, s2) } // u2 = u2+pd
+
             let rr := mulmod(0x02, sub(s2, s1), pd) // r = 2*(S2-S1)
             r1 := mulmod(rr, rr, pd) // X3 = R^2
 
             let v := mulmod(u1, i, pd) // V = U1*I
             let j2v := addmod(j, mulmod(0x02, v, pd), pd)
-            if lt(r1, j2v) {
-                r1 := add(pd, r1) // X3 = X3+pd
-            }
+            if lt(r1, j2v) { r1 := add(pd, r1) } // X3 = X3+pd
+
             r1 := sub(r1, j2v)
 
             // Y3 = r*(V-X3)-2*S1*J
             let s12j := mulmod(mulmod(0x02, s1, pd), j, pd)
 
-            if lt(v, r1) {
-                v := add(pd, v)
-            }
+            if lt(v, r1) { v := add(pd, v) }
             r2 := mulmod(rr, sub(v, r1), pd)
 
-            if lt(r2, s12j) {
-                r2 := add(pd, r2)
-            }
+            if lt(r2, s12j) { r2 := add(pd, r2) }
             r2 := sub(r2, s12j)
 
             // Z3 = ((Z1+Z2)^2-Z1Z1-Z2Z2)*H
             z1z1 := addmod(z1z1, z2z2, pd)
             j2v := mulmod(p3q3, p3q3, pd)
-            if lt(j2v, z1z1) {
-                j2v := add(pd, j2v)
-            }
+            if lt(j2v, z1z1) { j2v := add(pd, j2v) }
             r3 := mulmod(sub(j2v, z1z1), h, pd)
         }
         return (r1, r2, r3);
@@ -294,39 +239,28 @@ library LibSecp256r1 {
 
     // Point doubling on the modified jacobian coordinates
     // http://point-at-infinity.org/ecc/Prime_Curve_Modified_Jacobian_Coordinates.html
-    function _modifiedJacobianDouble(
-        uint x,
-        uint y,
-        uint z
-    ) internal pure returns (uint x3, uint y3, uint z3) {
+    function _modifiedJacobianDouble(uint256 x, uint256 y, uint256 z)
+        internal
+        pure
+        returns (uint256 x3, uint256 y3, uint256 z3)
+    {
         if (y == 0) return (0, 0, 0);
         assembly {
-            let
-                pd
-            := 0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFF
+            let pd := 0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFF
             let z2 := mulmod(z, z, pd)
-            let az4 := mulmod(
-                0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFC,
-                mulmod(z2, z2, pd),
-                pd
-            )
+            let az4 :=
+                mulmod(0xFFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFC, mulmod(z2, z2, pd), pd)
             let y2 := mulmod(y, y, pd)
             let s := mulmod(0x04, mulmod(x, y2, pd), pd)
             let u := mulmod(0x08, mulmod(y2, y2, pd), pd)
             let m := addmod(mulmod(0x03, mulmod(x, x, pd), pd), az4, pd)
             let twos := mulmod(0x02, s, pd)
             let m2 := mulmod(m, m, pd)
-            if lt(m2, twos) {
-                m2 := add(pd, m2)
-            }
+            if lt(m2, twos) { m2 := add(pd, m2) }
             x3 := sub(m2, twos)
-            if lt(s, x3) {
-                s := add(pd, s)
-            }
+            if lt(s, x3) { s := add(pd, s) }
             y3 := mulmod(m, sub(s, x3), pd)
-            if lt(y3, u) {
-                y3 := add(pd, y3)
-            }
+            if lt(y3, u) { y3 := add(pd, y3) }
             y3 := sub(y3, u)
             z3 := mulmod(0x02, mulmod(y, z, pd), pd)
         }
@@ -336,22 +270,16 @@ library LibSecp256r1 {
     // a^(p-1) = 1 mod p
     // a^(-1) ≅ a^(p-2) (mod p)
     // we then use the precompile bigModExp to compute a^(-1)
-    function _primemod(uint value, uint p) internal view returns (uint ret) {
+    function _primemod(uint256 value, uint256 p) internal view returns (uint256 ret) {
         ret = modexp(value, p - 2, p);
         return ret;
     }
 
     // Wrapper for built-in BigNumber_modexp (contract 0x5) as described here. https://github.com/ethereum/EIPs/pull/198
-    function modexp(
-        uint _base,
-        uint _exp,
-        uint _mod
-    ) internal view returns (uint ret) {
+    function modexp(uint256 _base, uint256 _exp, uint256 _mod) internal view returns (uint256 ret) {
         // bigModExp(_base, _exp, _mod);
         assembly {
-            if gt(_base, _mod) {
-                _base := mod(_base, _mod)
-            }
+            if gt(_base, _mod) { _base := mod(_base, _mod) }
             // Free memory pointer is always stored at 0x40
             let freemem := mload(0x40)
 
@@ -365,12 +293,8 @@ library LibSecp256r1 {
 
             let success := staticcall(14000, 0x5, freemem, 0xc0, freemem, 0x20)
             switch success
-            case 0 {
-                revert(0x0, 0x0)
-            }
-            default {
-                ret := mload(freemem)
-            }
+            case 0 { revert(0x0, 0x0) }
+            default { ret := mload(freemem) }
         }
     }
 }
